Fix infinite loop in q_sort
The slow fast pointer using for loop checks the condition before running the loop body, therefore an additional pointer "tail" was used to store the tail of the queue. The new implementation uses do ... while loop, which checks the condiiton after running the loop body, thus lessen the use of additional pointers.