Fix q_shuffle randomness
A new implementation of Fisher-Yates algorithm is added to improve
randomness of q_shuffle. The original implementation only moves the
random chosen node to tail of queue, which was proven not
an uniform distribution with Pearson's chi-squared test with p-value of 0.03451, which is
below chosen alpha = 0.05. The new implementation had been tested to
give a better performence with a p-value of 0.97.